requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
ase==3.21.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
bokeh==2.3.1
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
colorcet==2.0.6
cox==0.1.post3
cycler==0.10.0
dask==2021.4.1
datashader==0.12.1
datashape==0.5.2
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
distributed==2021.4.1
docopt==0.6.2
entrypoints==0.3
flatbuffers==1.12
fsspec==2021.4.0
gast==0.3.3
gin-config==0.4.0
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googledrivedownloader==0.4
grpcio==1.32.0
h5py==2.10.0
HeapDict==1.0.1
holoviews==1.14.3
idna==2.10
imageio==2.9.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
jsonpatch==1.28
jsonpickle==1.4.2
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
littleutils==0.2.2
llvmlite==0.35.0
locket==0.2.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
msgpack==1.0.2
multipledispatch==0.6.0
munch==2.5.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.5
numba==0.52.0
numexpr==2.7.3
numpy==1.19.4
oauthlib==3.1.0
ogb==1.2.4
opencv-python==4.4.0.46
opt-einsum==3.3.0
outdated==0.2.0
packaging==20.8
pandas==1.2.0
pandocfilters==1.4.3
panel==0.11.3
param==1.10.1
parso==0.7.1
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.6.0
py-cpuinfo==7.0.0
py3nvml==0.2.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyct==0.4.8
Pygments==2.7.3
pynndescent==0.5.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-louvain==0.15
pytz==2020.5
pyviz-comms==2.0.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
rdflib==5.0.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
sacred==0.8.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
sortedcontainers==2.3.0
tables==3.6.1
tblib==1.7.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.2
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tfrecord==1.12
threadpoolctl==2.1.0
tifffile==2021.4.8
toolz==0.11.1
torch==1.7.1+cu110
torch-geometric==1.6.3
torch-scatter==2.0.5
torchaudio==0.7.2
torchfile==0.1.0
torchvision==0.8.2+cu110
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
typing-extensions==3.7.4.3
umap-learn==0.5.1
urllib3==1.26.2
visdom==0.1.8.9
visdom-logger @ git+https://github.com/luizgh/visdom_logger.git@3c0afb18321ef65382d0baf8866772c6a62f15a2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xarray==0.17.0
xmltodict==0.12.0
zict==2.0.0
